LILY-2264: Limit features based on current payment plan

+++ lily/users/api/views.py
@@ -26,7 +26,7 @@ from templated_email import send_templated_mail
 
 from lily.utils.functions import post_intercom_event
 
-from lily.billing.functions import update_subscription
+from lily.utils.functions import has_required_tier
 
 from .utils import get_info_text_for_device
 from .serializers import TeamSerializer, LilyUserSerializer, LilyUserTokenSerializer, SessionSerializer
@@ -161,6 +161,9 @@ class LilyUserViewSet(viewsets.ModelViewSet):
         POST generates a new token
         DELETE removes the current token
         """
+        if not has_required_tier(2):
+            return Response(status=status.HTTP_403_FORBIDDEN)
+
         if request.method in ('DELETE', 'POST'):
             Token.objects.filter(user=request.user).delete()
             if request.method == 'DELETE':
@@ -353,7 +356,7 @@ class TwoFactorDevicesViewSet(viewsets.ViewSet):
             # Don't call super, since that only fires another query using self.get_object().
             self.perform_destroy(user_to_delete)
 
-            update_subscription(tenant, -1)
+            tenant.billing.update_subscription(-1)
 
             return Response(status=status.HTTP_204_NO_CONTENT)
         else:
